added ~/.humanism_c_history file. c() hits are recorded and searched first before looking in subdirs. This file is managed and monitored so that accidental hits are purged. The following common behavior causes a hit to NOT be recorded (or actually, for it to be removed for the history)
~/$ c filter
~/.didnt_want_to_go_here/somefilter$
 # whoops, we didn't want to go here. If user changes to a directory
 # not under the current path within 5 seconds the record is not recorded
 # in history
~/.didnt_want_to_go_here/somefilter$ c betterfilter
~/path/betterfilterhit$

This "feature" provides a pscyhological benifit that once the hit is
recorded with a specific filter this filter acts like a tag for the future.
From the example above "c filter" executed from anywhere will always land
you in ~/path/betterfilterhit.

This may also be considered a "bug" however and for this reason will remain
out of the main branch until the issue is resolve. The bug lays in that
the history takes priority. With the above example "filter" becomes a mental
tag but "path" becomes a hidden priority. So if you know you want to go to
a directory under or near the current using the filter "path" you would wind
up going to the hit in the history rather than something nearby.

There are several ways to resolve this "bug". The cleanest that would not
break current workflow of master branch c() is to check history only after
c() function fails to find the filter near by. But this then nearly negates
the ability for using filters as dynamic tags. The might be a compromise
and only searching the history hits if filter is not found in sub path but
before searching in parent paths. This is still weakens its function for
psychological tagging. Alternatively one could record the filter with the
hit path in the history file and only search on filters. Hmm