working off of the last commit message this changes things a little. We now record the search filter along with the directory hit. When c() is executed it searches for a match of the filter. This avoids collisions better. However there is still a problem with using filters to create tags when those filters are shared and common directory names or simply common filters one would want to use with a tree search. Consider a directory structure:
gitrepos/project1/.git
gitrepos/project1/images
gitrepos/project2/.git
gitrepos/project2/images

Sufficient global tags would be project1, project2. Bad global tags
would be git, project or images. If cwd is project1 we will
likely call ``c imag`` thereby creating the imag tag for
``gitrepos/project1/images``. This will cause a problem when cwd
is project2 and we want to ``c imag`` within that directory.
``c images`` will not have this problem because currently a filter
that validates with ``-d filter`` takes preference.

Alternatively we could only search global tags when a c() fails for
subtree and just before we resort to searching in parents. But this
would be a compromise. A collision could arrise when we want to go
to a generic tag, say 'gitrepo' when the cwd has a match. Because
tags are auto created from filters and c() calls it's not clear
how much of a problem this creates. However, for now, for the purpose
of experimentation, this is the compromise this include uses.