c() history tags now match case-insensitive.
Thoughts: having used this for a few days I think the behaivor
of searching under sub dir for filter, then searching history/tags
and then searching parent trees, works. However one option that
could be added is the option to keep the lcdrc tag list pure and
only add to it on request. so perhaps a cpush() function with no
arguments to add the last (current cwd) entry in the history to the
tags. To implement this feature the the push_history function should
utilize a seperate history file from the lcdrc tag list and a cpush()
function should be made to pull the last entry as a tag into lcdrc