Major. Removed minor features, added useful others.
c()'s functionality remains the same.

l() is gone but can be easily added with an alias:

    alias l="cascade_command 'ls -l --color'
    complete -o plusdirs -A file -F _cascade_completion l

In fact cascade_command provides the ability to utilize the search
and tag db features on any command that accepts directories as
arguments. For instance, use tag names or fuzzy searching on
your ide/editor such as sublime or atom:

    alias atom="cascade_command atom"
    complete -o plusdirs -A file -F _cascade_completion atom

This means in the same way that c <tag> <FiLtEr> <etc> takes
you to found or matched path, now this match would be passed
to the command given to cascade_command (the atom editor in
the example)

This revision also removes auto pruning of tags. This was a
feature that would attempt to detect when an auto-added tag was
a mistake. But I found this was never trustworthy and I resorted
to manually deleting auto-added tags with `cc d` on the rare
occasion an auto-add might create issues.

Also made some assumptions to reduce code. We assume:
* It is okay to have multiple tags with the same path...
  (was HUMANISM_C_TAG_UNIQUE option before)
* ...so long as the tag names most recently used are always prioritized
  (was HUMANISM_C_TAG_PRIORITIZE_RECENT)

All debug statements removed. I may regret this later but for now
it makes code more readable.