pass ps supported arguments to ps. issue #5
as requested this will pass ps arguments to ps. to determine what is a 
valid argument and if that argument requires an additional parameter it 
is tested with ps and if not error results then it assumes it is part of 
a valid ps argument, else assumes it is part of the filter. arguments 
have to be given first as arguments. 

this feature was requested in:
https://github.com/cyphunk/humanism.sh/issues/5